Refactor socket service script
- Move the socketService.js script to /services.

- The token cookie is used to verify the user instead of email.

- Install cookie dependency to read cookies from the socketService
script.

- Refactor SocketClient component to grab isAuthenticated instead of
user's email from the Redux store.
