+++ server/package.json
@@ -23,6 +23,7 @@
     "bcryptjs": "^2.4.3",
     "bull": "^3.12.1",
     "colors": "^1.4.0",
+    "cookie": "^0.4.0",
     "cookie-parser": "^1.4.4",
     "cors": "^2.8.5",
     "currency.js": "^1.2.2",
